+++ sdk/include/rs/core/lrs_image.h
@@ -24,7 +24,7 @@ namespace rs
                       image_interface::flag flags,
                       rs::utils::smart_ptr<metadata_interface> metadata);
             image_info query_info(void) const override;
-            uint64_t query_time_stamp(void) const override;
+            double query_time_stamp(void) const override;
             flag query_flags(void) const override;
             const void * query_data(void) const override;
             stream_type query_stream_type() const override;
